ml-docker-app/requirements.txt
fastapi
uvicorn
scikit-learn
xgboost
matplotlib
Flask==2.3.2
Pillow==9.2.0
joblib==1.2.0
numpy==1.24.3
pandas==1.5.3
xgboost==1.7.6
